Removed asm dir from repo
A couple functions matched
Vec2i_union struct for pos (since x and y seem to be referenced as word / half often)
Struggles with func_80027370

+++ include/data_symbols.h
@@ -261,7 +261,7 @@ extern UNK_TYPE D_800BE614;
 extern UNK_TYPE D_800BE618;
 extern uint16_t D_800BE62C;
 extern uint16_t D_800BE630;
-extern int32_t D_800BE638;
+extern int16_t D_800BE638;
 extern int16_t D_800BE63C;
 extern int16_t D_800BE668;
 extern uint16_t D_800BE66C;
